Migrate to Typescript the module routing/routes.tsx
